+++ js/admlteext.js
@@ -4,19 +4,24 @@ $(function() {
 	/*
     * Add collapse and remove events to boxes And save cookies
     */
-		$("[data-widget='collapse']").off("click");
-		$("[data-widget='collapse']").click(function() {
+		//$("[data-widget='collapse']").off("click");
+		$("[data-widget='collapse']").each(function() {
 			var box = $(this).parents(".box").first();
-			//Find the body and the footer
-			var bf = box.find(".box-body, .box-footer");
+			//console.log('found '+box.attr('id') + 'cookstate = ' + $.cookie(box.attr('id')+'_state') );
+				if($.cookie(box.attr('id')+'_state') == "hide"){
+					if (!box.hasClass("collapsed-box")) {
+                         box.addClass("collapsed-box");
+						 box.slideDown();
+					};
+			}
+		});
+		$("[data-widget='collapse']").on('click',function() {
+			var box = $(this).parents(".box").first();
+			//console.log('clicked '+box.attr('id') + 'cookstate = ' + $.cookie(box.attr('id')+'_state') );
 			if (!box.hasClass("collapsed-box")) {
-				box.addClass("collapsed-box");
 				$.cookie(box.attr('id')+'_state',"hide");
-				bf.slideUp();
 			} else {
-				box.removeClass("collapsed-box");
 				$.cookie(box.attr('id')+'_state',"show");
-				bf.slideDown();
 			}
 		});
 });