+++ CHANGELOG.md
@@ -8,3 +8,5 @@ CHANGELOG
 13.03.2015 v1.1.1 new Info-Box widget for AdminLte 2.0
            add with-border options to Box widget for AdminLte 2.0
            add ability to set left and right toolbar elements for box widget not only as string html code but also as array for yii\bootstrap\ButtonGroup $buttons
+25.09.2015 v1.1.2 Fix collapse remember functional for Box and Tiles,
+                  add new options for Box and Tile "collapseDefault" - if true, box will be collapsed by default